Local RTC Timer laeuft nun mit belibiegem Interval I2C Treiber benoetigt Fehlerbehandlung fuer Status = 0x00

+++ src/driver/local_rtc_driver.h
@@ -50,7 +50,7 @@ u8 local_rtc_timer_istimeup_u8(u8 time_reference, u8 time_interval);
  * @param time_interval
  * @return
  */
-u16 local_rtc_timer_istimeup_u16(u16 time_reference, u16 time_interval);
+u8 local_rtc_timer_istimeup_u16(u16 time_reference, u16 time_interval);
 
 /*!
  *
@@ -58,6 +58,6 @@ u16 local_rtc_timer_istimeup_u16(u16 time_reference, u16 time_interval);
  * @param time_interval
  * @return
  */
-u32 local_rtc_timer_istimeup_u32(u32 time_reference, u32 time_interval);
+u8 local_rtc_timer_istimeup_u32(u32 time_reference, u32 time_interval);
 
 #endif
